Rename any reference to 'regression analysis' to 'change detection'
The term 'regression analysis' has wider meaning in analytics and we tend to use it in context of scalability testing, too. It overlaps with meaning of 'regression' in continuous performance testing. The current implementation is looking for any significant changes over time, so we'll try to clarify the terminology here.