Add Action to post message to slack
The behavior of the action is similar to the existing GitHub actions,
and largely a cookie-cutter clone. (Some common code could be extracted
into a base class, but I resisted making unnecessary changes.)

Essentially, it accepts a Slack service token, a Slack channel ID, and
formatter, and posts the formatted message to the designated Slack
channel.

On the Slack side, the setup required includes:

1 Define and install a Slack "app" in your Slack workspace, using the
  slack CLI or https://api.slack.com/apps, then generate an app token
  with at least scopes "channel:join" and "chat:write". (Note that a
  workspace may require administrative review and approval to install an
  app.) The token should be a long alphanumeric string starting with
  "xoxb-".
2 Use the Slack UI (or API) to determine the ID of the channel you want
  to use. (Which looks something like "C00A000A0A".)
3 Use the https://api.slack.com/methods/conversations.join/test API
  tester with your token and the channel ID to "join" your app to the
  channel so it can post. (There are of course other ways, but this is
  probably the easiest, and provides a quick verification that the token
  and channel ID are good.)

Creating a `slack-channel-message` action requires two "config"
parameters and one "secrets" parameter:

1 The "channel" conf value
2 The Horreum "formatter" name
3 The "token" secret

The Action will always post a "mrkdwn" Slack message using the markdown
document generated by the Horreum formatter template. It will post using
the name of the Slack app (e.g., "Horreum Server").

Note that there's some incompatibility between GitHub and Slack markdown
syntax, specifically for links. Existing Horreum markdown templates for
GitHub use the `[text](url)` syntax whereas Slack uses `<url|text>`.
This suggests that Horreum may need to extend the templating classes to
conditionalize link formatting based on the action context.

For testing (and as an example), this PR proposes a new formatter
template for the `Test` class, using Slack link syntax.

Finally, I tried to add testing for this using the framework in the
`ActionServiceTest.java` module. The new test attempts to create a Slack
action, force it to execute, and validate failure (due to a "bad" fake
channel and token) by querying the action log. The code fails to find
any log entries either for the actual test ID or for test "-1", which
appears to be what `ActionServiceImpl.java` actually uses when logging
an error. Errors are however visible in the log output.

Resolves issue #838